chore: move eval compare script into evals
